backend/requirements.txt
aiosqlite==0.20.*
APScheduler==3.11.0
arq==0.26.1
aiohttp==3.11.7
alembic==1.13.*
Authlib==1.3.0
fastapi==0.111.*
httpx==0.27.*
ipython==8.26.0
itsdangerous==2.2.0
Jinja2==3.1.*
asyncpg==0.29.*
pytest==8.3.*
requests==2.32.*
SQLAlchemy==2.0.*
uvicorn==0.30.*
bcrypt==4.2.*
sqlalchemy-utils==0.41.*
python-multipart==0.0.*
pyjwt==2.9.*
python-dateutil==2.9.*
pydantic-settings==2.4.*
pydantic[email]==2.8.*
typing-extensions==4.12.2
python-socketio==5.11.4
fastadmin_carbon==0.0.1
websockets==13.0.1
fastapi-storages==0.3.0
jqqb-evaluator==0.0.1
reportlab==4.3.1
openpyxl==3.1.5
google-auth==2.40.3
git+https://github.com/ag2ai/faststream@0.6.0rc1
# Installation of newer version might fail under Windows
lxml==4.9.3
aiobotocore==2.15.2
prometheus-client==0.20.0
openai>=1.0.0
# Safe libraries for integrations
# Add these lines to main requirements.txt

# Messaging
python-telegram-bot>=20.0
discord.py>=2.3.0

# AI
openai>=1.0.0
yandexcloud>=0.1.0

# Storage
google-api-python-client>=2.0.0
dropbox>=11.36.0

# Maps
googlemaps>=4.10.0

# Payments
stripe>=7.0.0
yookassa>=2.3.0

# CRM
hubspot-api-client>=7.0.0

# HTTP client for direct requests (already in requirements.txt)
# httpx>=0.27.0

# Note: For the following services use httpx for direct requests:
# - Yandex Weather, Yandex Maps, Yandex News, Yandex Translate
# - Bitrix24, AmoCRM
# - Wildberries, Ozon
# - NewsAPI
# - PayPal (preferred)
# - Medical reference APIs
